moved modal into seperate component
