+++ src/components/pages/concert.css
@@ -63,112 +63,6 @@
   justify-content: center;
 }
 
-.Modal {
-  position: absolute;
-  top: 1em;
-  bottom: 1em;
-  right: 1em;
-  left: 1em;
-  background: var(--off-white);
-  display: grid;
-  grid-template-rows: 23% 1fr;
-  border: 5px solid var(--pink) ;
-
-  form {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: flex-end;
-    font-family: var(--header-font);
-
-    input {
-      width: 80%;
-      height: 2em;
-      text-align: center;
-      font-size: 1em;
-      border: none;
-      background: var(--dark-blue);
-      color: var(--pink);
-      margin: .4em;
-      font-family: var(--paragraph-font);
-    }
-
-    button {
-      margin: .2em;
-      padding: .2em;
-      width: 9em;
-      font-size: 1em;
-      font-family: var(--paragraph-font);
-      background: var(--dark-blue);
-      color: var(--pink);
-    }
-  }
-
-  ul {
-    overflow: scroll;
-    padding: 0;
-    margin: 0;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-
-    li {
-      width: 80%;
-      list-style: none;
-      display: flex;
-      align-items: center;
-      font-family: var(--paragraph-font);
-      background: var(--dark-blue);
-      color: var(--pink);
-      padding: .3em;
-      margin: 1em;
-
-      img {
-       height: 4em;
-      }
-
-      h2 {
-        font-size: 1em;
-        font-family: var(--paragraph-font);
-        padding: 1em;
-      }
-    }
-  }
-}
-
-.Exit {
-  position: absolute;
-  top: .5em;
-  right: .5em;
-  background: var(--dark-blue);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  font-size: 1.3em;
-  border: none;
-  color: var(--pink);
-  font-family: var(--paragraph-font);
-}
-
-.Shake {
-  animation: shake 1s ease-in-out;
-}
-
-@keyframes shake {
-  0% { transform: translate(1px, 1px) rotate(0deg); }
-  10% { transform: translate(-1px, -2px) rotate(-1deg); }
-  20% { transform: translate(-3px, 0px) rotate(1deg); }
-  30% { transform: translate(3px, 2px) rotate(0deg); }
-  40% { transform: translate(1px, -1px) rotate(1deg); }
-  50% { transform: translate(-1px, 2px) rotate(-1deg); }
-  60% { transform: translate(-3px, 1px) rotate(0deg); }
-  70% { transform: translate(3px, 1px) rotate(-1deg); }
-  80% { transform: translate(-1px, -1px) rotate(1deg); }
-  90% { transform: translate(1px, 2px) rotate(0deg); }
-  100% { transform: translate(1px, -2px) rotate(-1deg); }
-}
-
-
 .Playlist {
   width: 85%;
   display: flex;
@@ -213,10 +107,6 @@
 }
 
 @media screen and (min-width: 600px) {
-  .Modal {
-    right: 20%;
-    left: 20%;
-  }
   .Playlist {
     width: 60%;
   }
