+++ src/hooks/playlistState.js
@@ -1,4 +1,4 @@
-import { useEmitEvent } from '../../socket';
+import { useEmitEvent } from '../socket';
 
 export default function usePlaylistEmitters() {
   const addLink = useEmitEvent('ADD_LINK');
